Create Morse Characters
just put the name of the message as camelcase (ex. messageMorse) and then fill the bracket in with the index of array (ex. messageMorse[0],messageMorse[1], ...) Ik it's painful to do blame gamemaker. and pls keep the comment of the letter so it is easier to read and debug if needed thx.